Moved js code from index.html to global.js
I didn't like how each file needed to know what variables were global and how I
needed to maintain a .jshintrc. Instead I have no global variables and
import/require everything needed per file.

This also made circular dependencies more obvious so I had to add some
callbacks to deal with resulting errors
