+++ .jshintrc
@@ -61,7 +61,7 @@
 	"couch"         : false,    // Enable globals exposed by CouchDB.
 	"devel"         : false,    // Allow development statements e.g. `console.log();`.
 	"dojo"          : false,    // Enable globals exposed by Dojo Toolkit.
-	"jquery"        : true ,    // Enable globals exposed by jQuery JavaScript library.
+	"jquery"        : false,    // Enable globals exposed by jQuery JavaScript library.
 	"mootools"      : false,    // Enable globals exposed by MooTools JavaScript framework.
 	"node"          : true,     // Enable globals available when code is running inside of the NodeJS runtime environment.
 	"nonstandard"   : false,    // Define non-standard but widely adopted globals such as escape and unescape.
