+++ plugins/Hosting/index.html
@@ -66,42 +66,7 @@
 		</div>
 
 		<!-- JS -->
-		<script>
-			// Library for communicating with Sia-UI
-			const IPCRenderer = require('electron').ipcRenderer;
-			// Jquery
-			const $ = require('jquery');
-
-			// Host settings manager
-			var Host = require('./js/host.js')
-			// Lifecycle manager
-			var Lifecycle = require('./js/lifecycle.js');
-
-			// Siad wrapper/manager
-			var Siad = require('sia.js')
-			// Make sure Siad settings are in sync with the rest of the UI's
-			IPCRenderer.sendToHost('config', {key: 'siad'}, 'siadsettings');
-			IPCRenderer.on('siadsettings', function(event, settings) {
-				Siad.configure(settings);
-			});
-			// Slight modification to Siad wrapper for standard error handling
-			Siad.apiCall = function(callObj, callback) {
-				Siad.call(callObj, function(err, result) {
-					if (err) {
-						console.error(callObj, err);
-						Lifecycle.notify(err.toString(), 'error');
-					} else {
-						callback(result);
-					}
-				});
-			}
-
-			// Expose start and stop methods to global namespace so they're
-			// automatically called by the general UI
-			var start = Lifecycle.update;
-			var stop = Lifecycle.stop;
-		</script>
-		<script src='./js/buttons.js'></script>
+		<script src='./js/global.js'></script>
 	</body>
 </html>
